feat: optimize game by sponsoring creating the game
