+++ app/src/helpers/sponsorAndSignTransactionWithApi.ts
@@ -1,0 +1,14 @@
+import {fromBase64, toBase64} from "@mysten/bcs";
+import { SuiClient } from "@mysten/sui/src/client";
+import { Transaction} from "@mysten/sui/src/transactions";
+
+interface Args {
+	tx: Transaction;
+	suiClient: SuiClient;
+	sender: string;
+}
+
+export default function sponsorAndSignTransactionWithApi<T>({ tx, suiClient, sender }: Args) {
+
+
+}
