+++ app/src/hooks/useMakeMoveInBlackjackGame.ts
@@ -7,7 +7,7 @@ import axios from "axios";
 import { useEnokiFlow, useZkLogin } from "@mysten/enoki/react";
 import { useSui } from "./useSui";
 import { doHit, doStand } from "@/__generated__/blackjack/single_player_blackjack";
-import { fromBase64, toBase64 } from "@mysten/bcs";
+import useSponsoredTransaction from "@/hooks/useSponsoredTransaction";
 
 interface HandleHitOrStandProps {
   move: "hit" | "stand";
@@ -28,8 +28,7 @@ interface OnRequestMoveSuccessProps {
 }
 
 export const useMakeMoveInBlackjackGame = () => {
-  const { suiClient } = useSui();
-  const enokiFlow = useEnokiFlow();
+  const { sponsorAndSignTransaction } = useSponsoredTransaction();
   const { address } = useZkLogin();
   const [isMoveLoading, setIsMoveLoading] = useState(false);
 
@@ -64,69 +63,7 @@ export const useMakeMoveInBlackjackGame = () => {
           tx.pure.address(process.env.NEXT_PUBLIC_ADMIN_ADDRESS!)
         );
 
-        const txBytes = await tx.build({
-          client: suiClient,
-          onlyTransactionKind: true,
-        });
-
-        // Step 2: Send TxBytes to the backend for sponsorship
-        const sponsorResponse = await fetch("/api/sponsor", {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({
-            transactionKindBytes: toBase64(txBytes),
-            sender: address,
-          }),
-        });
-
-        if (!sponsorResponse.ok) {
-          throw new Error("Failed to sponsor transaction");
-        }
-
-        const { bytes: sponsoredBytes, digest: txDigest } =
-          await sponsorResponse.json();
-
-        // Step 3: User signs the sponsored TxBytes
-        const signer = await enokiFlow.getKeypair({
-          network: process.env.NEXT_PUBLIC_SUI_NETWORK_NAME! as
-            | "mainnet"
-            | "testnet",
-        });
-
-        const { signature } = await signer.signTransaction(
-          fromBase64(sponsoredBytes)
-        );
-
-        // Step 4: Send signed TxBytes and txDigest back to the backend for execution
-        const executeResponse = await fetch("/api/execute", {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({
-            digest: txDigest,
-            signature: signature,
-          }),
-        });
-
-        if (!executeResponse.ok) {
-          throw new Error("Failed to execute transaction");
-        }
-
-        const { digest: executedDigest } = await executeResponse.json();
-
-        // Step 5: Wait for transaction confirmation
-        await suiClient.waitForTransaction({
-          digest: executedDigest,
-          timeout: 10_000,
-        });
-
-        const transactionResult = await suiClient.getTransactionBlock({
-          digest: executedDigest,
-          options: {
-            showEffects: true,
-            showObjectChanges: true,
-            showEvents: true,
-          },
-        });
+        const transactionResult = await sponsorAndSignTransaction(tx, address!);
 
         if (transactionResult.effects?.status?.status !== "success") {
           throw new Error("Transaction failed");
